for server to run

+++ main.py
@@ -1,7 +1,7 @@
 import pandas as pd  # Import the pandas library for data manipulation
 import lyapunov_final as lyap  # Import the lyapunov_final module
 import numpy as np  # Import the numpy library for numerical operations
-import truncate as tr  # Import the truncate module
+import process_data as tr  # Import the truncate module
 import state_space as ss  # Import the state_space module
 import psd as psd  # Import the psd module
 import filename_generation as fg  # Import the filename_generation module
@@ -22,7 +22,8 @@ def main():
     plotting_final = 300  # Ending point for plotting
     tolerance = 0.001  # Tolerance for truncation
 
-    df = pd.read_csv("1_raw_data/running_info.csv")  # Read the CSV file into a pandas DataFrame
+    df = pd.read_csv("1_raw_data/running_info_all.csv")  # Read the CSV file into a pandas DataFrame
+    filtered_df = pd.read_csv("1_raw_data/running_info.csv")  # Read the CSV file into a pandas DataFrame
     data = df.values  # Convert the DataFrame to a numpy array
     dates = df['Date'].values  # Get the dates from the 'Date' column
     dates = np.unique(dates)  # Get the unique dates
@@ -38,48 +39,51 @@ def main():
         for terrain in terrains.get(date):
             trials.get(date).update({terrain: np.unique(data[(data[:, 0] == date) & (data[:, 1] == terrain)][:, 2])})  # Get the unique trials for each date and terrain
     for date in dates:
+        if date !="Simulation":
+            continue
         for terrain in terrains.get(date):
-            if("terrain" in terrain):
-                continue
             exponents = {}  # Initialize an empty dictionary to store Lyapunov exponents
             psds = {}  # Initialize an empty dictionary to store power spectral densities
             forward_velocities={} # Initialize an empty dictionary to store forward velocities
             for trial in trials.get(date).get(terrain):
                 trial_data=[date,terrain,trial]
-                print("Running for date:", date, ", terrain:", terrain, ", trial:", trial)
+                print("\nRunning for date:", date, ", terrain:", terrain, ", trial:", trial)
                 row = df[(df['Date'] == date) & (df['Terrain'] == terrain) & (df['Trial'] == trial)]  # Get the row corresponding to the current date, terrain, and trial
                 row_array = row.values[0]  # Convert the row to a numpy array
                 tr_start = row_array[3]  # Get the start time for truncation
                 tr_end = row_array[4]  # Get the end time for truncation
-                tr.retruncate(date, terrain, trial, tr_start, tr_end)  # Call the retruncate function from the truncate module
+                tr.truncate(date, terrain, trial, tr_start, tr_end)  # Call the retruncate function from the truncate module
                 tr.main(date, terrain, trial, tolerance)  # Call the main function from the truncate module
                 
-
                 info = pd.read_csv(fg.filename_clean_data(date, terrain, trial))  # Read the cleaned data from the CSV file
                 delta_t = np.mean(np.diff(info['timestamp'])) # Calculate the mean time step
 
                 ss.main(date, terrain, trial, delta_t,forward_velocities,trial_data)  # Call the main function from the state_space module
-                lyap.exponent(tau, m, min_steps, plotting_0, plotting_final,
+
+                
+                if ((filtered_df['Date'] == date) & (filtered_df['Terrain'] == terrain) & (filtered_df['Trial'] == trial)).any():
+                    lyap.exponent(tau, m, min_steps, plotting_0, plotting_final,
                               delta_t, force_minsteps, exponents, date, terrain, trial,trial_data)  # Call the exponent function from the lyapunov_final module
+                else:
+                    trial_data.append(np.nan)
+                
                 psd.main(psds, date, terrain, trial)  # Call the main function from the psd module
-                print("Trial data:",trial_data)
+                print("Trial Data:", trial_data)
                 data_store.append(trial_data)
 
             lyap.plot_exponents(exponents, date, terrain, 0)  # Call the plot_exponents function from the lyapunov_final module
-            psd.plot_psd(psds, date, terrain, 0)  # Call the plot_psd function from the psd module
-            ss.plot_forward_velocities(forward_velocities, date, terrain, 0)
+            ss.plot_forward_velocities(forward_velocities, date, terrain, 1)
+            ss.plot_forward_velocities(forward_velocities, date, terrain, 2)
 
             exponents_df = pd.DataFrame.from_dict(exponents, orient='index')
             exponents_df.to_csv(fg.filename_big(date, terrain, 0) + "exponents.csv")  # Save the DataFrame to a CSV file
 
-            psds_df = pd.DataFrame.from_dict(psds, orient='index')
-            psds_df.to_csv(fg.filename_big(date, terrain, 0) + "psds.csv")  # Save the DataFrame to a CSV file
-
             forward_velocities_df = pd.DataFrame.from_dict(forward_velocities, orient='index')
             forward_velocities_df.to_csv(fg.filename_big(date, terrain, 0) + "forward_velocities.csv")
 
+
     data_store_df = pd.DataFrame(data_store, columns=['Date', 'Terrain', 'Trial','Forward Velocities' ,'Exponent'])
-    data_store_df.to_csv("3_results/data.csv", index=False)  # Save the DataFrame to a CSV file
+    data_store_df.to_csv("3_results/velocities.csv", index=False)  # Save the DataFrame to a CSV file
